Change doc TODO comments into visible notes
Previously the TODO markers were only visible in the source. Now
they show up in the generated documentation, and can also be found
using the mkdocs search function. Every to-do item now has two
markers: TODO and TODO_XX, where XX are the initials of the assignee.
It is thus possible to search the generated documents both for
all to-dos and for those assigned to one or more specific people.